add tests and split course and grades test to separate files
